% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMonTools.R
\docType{package}
\name{BioMonTools-package}
\alias{BioMonTools}
\alias{BioMonTools-package}
\title{BioMonTools: Biomonitoring and Bioassessment Calculations}
\description{
BioMonTools provides functions to aid the data analysis of bioassessment and
biomonitoring data.  Suite of functions and tools for metric calculation and
scoring for multi-metric indices and related data manipulation.
}
\details{
An aid for manipulating data associated with biomonitoring and bioassessment.
Calculations include metric calculation, marking of excluded taxa,
subsampling, and multimetric index calculation.  Targeted communities are
benthic macroinvertebrates, fish, periphyton, and coral.  As described in the
Revised Rapid Bioassessment Protocols (Barbour et al. 1999)
<https://archive.epa.gov/water/archive/web/html/index-14.html>.

This software program is preliminary or provisional and is subject
to revision. This software program is for testing only, no warranty,
expressed or implied, is made as to the accuracy and functioning of the
program and related program material nor shall the fact of distribution
constitute any such warranty, and no responsibility is assumed in connection
therewith. This software is provided 'AS IS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leppott/BioMonTools}
  \item Report bugs at \url{https://github.com/leppott/BioMonTools/issues}
}

}
\author{
\strong{Maintainer}: Erik W. Leppo \email{Erik.Leppo@tetratech.com} (\href{https://orcid.org/0000-0001-8026-7474}{ORCID})

Other contributors:
\itemize{
  \item Jen Stamp [contributor]
  \item John van Sickles [contributor]
  \item Ben Block [contributor]
}

}
