% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_algae_names_official}
\alias{data_algae_names_official}
\title{Algae Taxa Names, Official}
\format{
A data frame with 3271 observations on the following 8 variables.

  \describe{
    \item{\code{Taxon}}{a character vector}
    \item{\code{Issues}}{a character vector}
    \item{\samp{Number of occurrences}}{a numeric vector}
    \item{\samp{modified BCG}}{a character vector}
    \item{\code{OTU_BCG}}{a character vector}
    \item{\code{Notes}}{a character vector}
    \item{\code{OTU_Carlisle}}{a character vector}
    \item{\code{OTU_Mertens}}{a character vector}
  }
}
\source{
example data
}
\usage{
data_algae_names_official
}
\description{
Translation table for algae taxa.
}
\keyword{datasets}
