% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_algae_names_user}
\alias{data_algae_names_user}
\title{Algae Taxa Names, User}
\format{
A data frame with 3271 observations on the following variable.

 \describe{
   \item{\code{Taxon}}{a character vector}
  }
}
\source{
example data
}
\usage{
data_algae_names_user
}
\description{
Example user taxa names
}
\keyword{datasets}
