% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_benthos_MBSS}
\alias{data_benthos_MBSS}
\title{Benthic macroinvertebrate taxa data; MBSS}
\format{
A data frame with 5,666 observations on the following 40 variables.

 \describe{
  \item{\code{INDEX_NAME}}{a character vector}
  \item{\code{SAMPLEID}}{a character vector}
  \item{\code{DATE}}{a character vector}
  \item{\code{TAXAID}}{a character vector}
  \item{\code{N_TAXA}}{a numeric vector, count}
  \item{\code{N_GRIDS}}{a numeric vector, number of grids in subsample
  (max = 30)}
  \item{\code{EXCLUDE}}{a character vector, whether taxon should be excluded
  from taxa richness metrics}
  \item{\code{INDEX_CLASS}}{a character vector, index region}
  \item{\code{Phylum}}{a character vector}
  \item{\code{Class}}{a character vector}
  \item{\code{Order}}{a character vector}
  \item{\code{Family}}{a character vector}
  \item{\code{Genus}}{a character vector}
  \item{\code{Other_Taxa}}{a character vector}
  \item{\code{Tribe}}{a character vector}
  \item{\code{FFG}}{a character vector}
  \item{\code{FAM_TV}}{a numeric vector}
  \item{\code{Habit}}{a character vector}
  \item{\code{TOLVAL}}{a numeric vector}
  \item{\code{TOLVAL2}}{a numeric vector}
  \item{\code{UFC}}{a numeric vector}
  \item{\code{UFC_Comment}}{a character vector}
  \item{\code{SUBPHYLUM}}{a character vector}
  \item{\code{SUBCLASS}}{a character vector}
  \item{\code{INFRAORDER}}{a character vector}
  \item{\code{SUBFAMILY}}{a character vector}
  \item{\code{LIFE_CYCLE}}{a character vector}
  \item{\code{BCG_ATTR}}{a character vector}
  \item{\code{THERMAL_INDICATOR}}{a character vector}
  \item{\code{LONGLIVED}}{a character vector}
  \item{\code{NOTEWORTHY}}{a character vector}
  \item{\code{FFG2}}{a character vector}
  \item{\code{HABITAT}}{a character vector}
  \item{\code{ELEVATION_ATTR}}{a character vector}
  \item{\code{GRADIENT_ATTR}}{a character vector}
  \item{\code{WSAREA_ATTR }}{a character vector}
  \item{\code{HABSTRUCT}}{a character vector}
  \item{\code{BCG_ATTR2}}{a character vector}
  \item{\code{NONTARGET}}{a logical vector}
  \item{\code{AIRBREATHER}}{a logical vector}
 }
}
\source{
example data from MBSS
}
\usage{
data_benthos_MBSS
}
\description{
A data set with example benthic macroinvertebrate data.
Calculate metrics then statistics.  Data from MBSS.
}
\keyword{datasets}
