% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_benthos_PacNW}
\alias{data_benthos_PacNW}
\title{Benthic macroinvertebrate taxa data; Pacific Northwest}
\format{
A data frame with 598 observations on the following 29 variables.
\describe{
\item{\code{INDEX_NAME}}{a character vector}
\item{\code{Index_Region}}{a character vector}
\item{\code{SITE_TYPE}}{a character vector}
\item{\code{SampleID}}{a character vector}
\item{\code{TaxaID}}{a character vector}
\item{\code{N_TAXA}}{a numeric vector}
\item{\code{Exclude}}{a logical vector}
\item{\code{NonTarget}}{a logical vector}
\item{\code{Phylum}}{a character vector}
\item{\code{Class}}{a character vector}
\item{\code{Order}}{a character vector}
\item{\code{Family}}{a character vector}
\item{\code{Subfamily}}{a character vector}
\item{\code{Tribe}}{a character vector}
\item{\code{Genus}}{a character vector}
\item{\code{BCG_Attr}}{a numeric vector}
\item{\code{Thermal_Indicator}}{a character vector}
\item{\code{FFG}}{a character vector}
\item{\code{Clinger}}{a character vector}
\item{\code{LongLived}}{a logical vector}
\item{\code{Noteworthy}}{a logical vector}
\item{\code{Habitat}}{a character vector}
\item{\code{SubPhylum}}{a logical vector}
\item{\code{InfraOrder}}{a logical vector}
\item{\code{Habit}}{a logical vector}
\item{\code{Life_Cycle}}{a logical vector}
\item{\code{TolVal}}{a logical vector}
\item{\code{FFG2}}{a logical vector}
\item{\code{TolVal2}}{a logical vector}
}
}
\source{
example data
}
\usage{
data_benthos_PacNW
}
\description{
A dataset with example taxa data and attributes for calculating
metric values.
}
\keyword{datasets}
