% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bio2rarify}
\alias{data_bio2rarify}
\title{rarify example data}
\format{
A data frame with 223 rows and 28 variables:
\describe{
   \item{SampleID}{Sample ID}
   \item{TaxaID}{unique taxonomic identifier}
   \item{N_Taxa}{number of individuals in sample}
}
}
\source{
example data
}
\usage{
data_bio2rarify
}
\description{
A dataset with example benthic macroinvertebrate data
(600 count) to be used with the rarify function.  Includes 12 samples.
}
\keyword{datasets}
