% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_coral_bcg_metric_dev}
\alias{data_coral_bcg_metric_dev}
\title{Coral taxa data; Florida BCG.}
\format{
A data frame with 2138 observations on the following 25 variables.

  \describe{
  \item{\code{DataSource}}{a character vector}
  \item{\code{SampleID}}{a character vector}
  \item{\code{TotTranLngth_m}}{a numeric vector}
  \item{\code{SampDate}}{a Date}
  \item{\code{TAXAID}}{a character vector}
  \item{\code{CommonName}}{a character vector}
  \item{\code{Juvenile}}{a logical vector}
  \item{\code{DiamMax_cm}}{a numeric vector}
  \item{\code{DiamPerp_cm}}{a numeric vector}
  \item{\code{Height_cm}}{a numeric vector}
  \item{\code{TotMort_pct}}{a numeric vector}
  \item{\code{BCG_ATTR}}{a character vector}
  \item{\code{Weedy}}{a character vector}
  \item{\code{LRBC}}{a logical vector}
  \item{\code{MorphConvFact}}{a numeric vector}
  \item{\code{Phylum}}{a character vector}
  \item{\code{Class}}{a character vector}
  \item{\code{SubClass}}{a character vector}
  \item{\code{Order}}{a character vector}
  \item{\code{Family}}{a character vector}
  \item{\code{Genus}}{a character vector}
  \item{\code{SubGenus}}{a character vector}
  \item{\code{Species}}{a character vector}
  \item{\code{INDEX_NAME}}{a character vector}
  \item{\code{INDEX_CLASS}}{a character vector}
}
}
\source{
example coral data from Florida BCG
}
\usage{
data_coral_bcg_metric_dev
}
\description{
A data set with example coral data.
Calculate metrics.  Data from Florida BCG providers.
}
\keyword{datasets}
