% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_coral_bcg_metric_qc}
\alias{data_coral_bcg_metric_qc}
\title{Coral metric value data; Florida BCG.}
\format{
A data frame with 100 observations on the following 19 variables.

  \describe{
  \item{\code{SAMPLEID}}{a character vector}
  \item{\code{INDEX_NAME}}{a character vector}
  \item{\code{INDEX_CLASS}}{a character vector}
  \item{\code{transect_area_m2}}{a numeric vector}
  \item{\code{ncol_total}}{a numeric vector}
  \item{\code{lcol_total}}{a numeric vector}
  \item{\code{nt_total}}{a numeric vector}
  \item{\code{ncol_Acropora}}{a numeric vector}
  \item{\code{ncol_AcroOrbi_m2}}{a numeric vector}
  \item{\code{pcol_Acropora}}{a numeric vector}
  \item{\code{nt_BCG_att123}}{a numeric vector}
  \item{\code{nt_BCG_att1234}}{a numeric vector}
  \item{\code{nt_BCG_att5}}{a numeric vector}
  \item{\code{pt_BCG_att5}}{a numeric vector}
  \item{\code{LCSA3D_samp_m2}}{a numeric vector}
  \item{\code{LCSA3D_BCG_att1234_m2}}{a numeric vector}
  \item{\code{LCSA3D_LRBC_m2}}{a numeric vector}
  \item{\code{ncol_SmallWeedy}}{a numeric vector}
  \item{\code{pcol_SmallWeedy}}{a numeric vector}
}
}
\source{
example coral metric results from Florida BCG
}
\usage{
data_coral_bcg_metric_qc
}
\description{
A data set with coral metric value data.
Used to compare to metric value calculations.  Data from Florida BCG providers.
}
\keyword{datasets}
