% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_metval_scmb_ibi}
\alias{data_metval_scmb_ibi}
\title{data_metval_scmb_ibi}
\format{
An object of class \code{data.frame} with 20 rows and 13 columns.
}
\source{
example data
}
\usage{
data_metval_scmb_ibi
}
\description{
Example data metrics

 @format A data frame with 20 observations on the following 13 variables.

  \describe{
    \item{\code{INDEX_NAME}}{a character vector}
    \item{\code{INDEX_REGION}}{a character vector}
    \item{\code{SampID}}{a character vector}
    \item{\code{nt_total}}{a numeric vector}
    \item{\code{nt_Mol}}{a numeric vector}
    \item{\code{ni_Noto}}{a numeric vector}
    \item{\code{pi_intol}}{a numeric vector}
    \item{\code{qc_nt_total}}{a numeric vector}
    \item{\code{qc_nt_Mol}}{a numeric vector}
    \item{\code{qc_ni_Noto}}{a numeric vector}
    \item{\code{qc_pi_intol}}{a numeric vector}
    \item{\code{qc_sum}}{a numeric vector}
    \item{\code{qc_nar}}{a character vector}
  }
}
\keyword{datasets}
