% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mmi_dev}
\alias{data_mmi_dev}
\title{Metric data for metric stats for mmi development}
\format{
A data frame with 10,574 observations on the following 10 variables.

 \describe{
   \item{\code{Class}}{a character vector}
   \item{\code{Ref_v1}}{a character vector}
   \item{\code{CalVal_Class4}}{a character vector}
   \item{\code{Unique_ID}}{a character vector}
   \item{\code{BenSampID}}{a character vector}
   \item{\code{CollDate}}{a character vector}
   \item{\code{CollMeth}}{a character vector}
   \item{\code{TaxaID}}{a character vector}
   \item{\code{Individuals}}{a numeric vector}
   \item{\code{Exclude}}{a logical vector}
 }
}
\source{
example data
}
\usage{
data_mmi_dev
}
\description{
A data set with example benthic macroinvertebrate data.
Calculate metrics then statistics.
}
\keyword{datasets}
