% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_Taxa_MA}
\alias{data_Taxa_MA}
\title{Estuary taxa data}
\format{
A data frame with 2,675 observations on the following 15 variables.
\describe{
 \item{\code{estuary}}{a factor with levels \code{BOSTON HARBOR} \code{BUZZARDS BAY} \code{CAPE COD BAY} \code{MASSACHUSETTS BAY} \code{WAQUOIT BAY}}
 \item{\code{CommonName}}{a factor with levels \code{ALEWIFE} \code{AMERICAN EEL} \code{AMERICAN LOBSTER} \code{AMERICAN PLAICE} \code{AMERICAN SAND LANCE} \code{AMERICAN SHAD} \code{ATLANTIC COD} \code{ATLANTIC CROAKER} \code{ATLANTIC HERRING} \code{ATLANTIC MACKEREL} \code{ATLANTIC MENHADEN} \code{ATLANTIC ROCK CRAB} \code{ATLANTIC SALMON} \code{ATLANTIC STINGRAY} \code{ATLANTIC STURGEON} \code{ATLANTIC TOMCOD} \code{BAY ANCHOVY} \code{BAY SCALLOP} \code{BLACK DRUM} \code{BLACK SEA BASS} \code{BLUE CRAB} \code{BLUE MUSSEL} \code{BLUEBACK HERRING} \code{BLUEFISH} \code{BROWN SHRIMP} \code{BUTTERFISH} \code{CHANNEL CATFISH} \code{COWNOSE RAY} \code{CUNNER} \code{DAGGERBLADE GRASS SHRIMP} \code{EASTERN OYSTER} \code{FOURSPINE STICKLEBACK} \code{GOBIES} \code{GREEN CRAB} \code{GREEN SEA URCHIN} \code{GRUBBY} \code{HADDOCK} \code{HOGCHOKER} \code{JONAH CRAB} \code{KILLIFISHES} \code{LONGHORN SCULPIN} \code{MULLETS} \code{MUMMICHOG} \code{NINESPINE STICKLEBACK} \code{NORTHERN KINGFISH} \code{NORTHERN PIPEFISH} \code{NORTHERN SEAROBIN} \code{NORTHERN SHRIMP} \code{OCEAN POUT} \code{OYSTER TOADFISH} \code{PINFISH} \code{POLLOCK} \code{QUAHOG} \code{RAINBOW SMELT} \code{RED DRUM} \code{RED HAKE} \code{ROCK GUNNEL} \code{SCUP} \code{SEA SCALLOP} \code{SEVENSPINE BAY SHRIMP} \code{SHEEPSHEAD MINNOW} \code{SHORTHORN SCULPIN} \code{SHORTNOSE STURGEON} \code{SILVER HAKE} \code{SILVERSIDES} \code{SKATES} \code{SMOOTH FLOUNDER} \code{SOFTSHELL CLAM} \code{SPINY DOGFISH} \code{SPOT} \code{SPOTTED SEATROUT} \code{STRIPED BASS} \code{SUMMER FLOUNDER} \code{TAUTOG} \code{THREESPINE STICKLEBACK} \code{WEAKFISH} \code{WHITE HAKE} \code{WHITE PERCH} \code{WINDOWPANE FLOUNDER} \code{WINTER FLOUNDER} \code{YELLOW PERCH} \code{YELLOWTAIL FLOUNDER}}
 \item{\code{LifeStage}}{a factor with levels \code{ADULTS} \code{EGGS} \code{JUVENILES} \code{LARVAE} \code{MATING} \code{PARTURITION} \code{SPAWNING}}
 \item{\code{SalZone}}{a factor with levels \code{>25 ppt} \code{0.5-25 ppt}}
 \item{\code{Winter}}{a numeric vector}
 \item{\code{Spring}}{a numeric vector}
 \item{\code{Summer}}{a numeric vector}
 \item{\code{Fall}}{a numeric vector}
 \item{\code{All}}{a numeric vector}
 \item{\code{TaxaName}}{Taxa Names for mapping}
 \item{\code{State}}{a factor with levels \code{MA}}
 \item{\code{Latitude}}{a numeric vector}
 \item{\code{Longitude}}{a numeric vector}
 \item{\code{Count}}{a numeric vector}
 \item{\code{PctDensity}}{a numeric vector}
}
}
\source{
example data
}
\usage{
data_Taxa_MA
}
\description{
A dataset with example fish taxa data and locations for mapping.
}
\keyword{datasets}
