% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_taxa_names_issues}
\alias{data_taxa_names_issues}
\title{Taxa name data frame with issues}
\format{
A data frame with 215 observations on the following 3 variables.

 \describe{
    \item{\code{Source}}{a character vector}
    \item{\code{Issue}}{a character vector}
    \item{\code{FinalID}}{a character vector}
  }
}
\source{
example data
}
\usage{
data_taxa_names_issues
}
\description{
A data frame with problems to use as example data
}
\keyword{datasets}
