% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_values.R
\name{metric.values.algae}
\alias{metric.values.algae}
\title{Calculate metric values, Algae}
\usage{
metric.values.algae(
  myDF,
  MetricNames = NULL,
  boo.Adjust = FALSE,
  cols2keep = NULL
)
}
\arguments{
\item{myDF}{Data frame of taxa.}

\item{MetricNames}{Optional vector of metric names to be returned.}

\item{boo.Adjust}{Optional boolean value on whether to perform adjustments of
values prior to scoring.  Default = FALSE but may be TRUE for certain
metrics.}

\item{cols2keep}{Column names of fun.DF to retain in the output.  Uses
column names.}
}
\value{
Data frame
}
\description{
Subfunction of metric.values for use with Algae.
}
\details{
For internal use only.  Called from metric.values().
}
\keyword{internal}
