% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_values.R
\name{metric.values.coral}
\alias{metric.values.coral}
\title{Calculate metric values, coral}
\usage{
metric.values.coral(
  myDF,
  MetricNames = NULL,
  boo.Adjust = FALSE,
  cols2keep = NULL,
  MetricSort = NA,
  boo.Shiny = FALSE,
  verbose
)
}
\arguments{
\item{myDF}{Data frame of taxa.}

\item{MetricNames}{Optional vector of metric names to be returned.}

\item{boo.Adjust}{Optional boolean value on whether to perform adjustments of
values prior to scoring.  Default = FALSE but may be TRUE for certain
metrics.}

\item{cols2keep}{Column names of fun.DF to retain in the output.  Uses
column names.}

\item{MetricSort}{How metric names should be sort; NA = as is
, AZ = alphabetical.  Default = NULL.}

\item{boo.Shiny}{Boolean value for if the function is accessed via Shiny.
Default = FALSE.}

\item{verbose}{Include messages to track progress.  Default = FALSE}
}
\value{
Data frame
}
\description{
Subfunction of metric.values for use with coral
}
\details{
For internal use only.  Called from metric.values().
}
\keyword{internal}
