% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa.R
\name{qc_taxa}
\alias{qc_taxa}
\title{Quality Control Check on User Data Against Master Taxa List}
\usage{
qc_taxa(
  DF_User,
  DF_Official = NULL,
  fun.Community = NULL,
  useOfficialTaxaInfo = "only_Official"
)
}
\arguments{
\item{DF_User}{User taxa data.}

\item{DF_Official}{Official master taxa list.  Can be a local file or
from a URL.
Default is NULL.  A NULL value will use the official online files.}

\item{fun.Community}{Community name for which to compare the master taxa list
(bugs or fish).}

\item{useOfficialTaxaInfo}{Select how to handle new/different taxa.
See 'Details' for more information.
Valid values are "only_Official", "only_user", "add_new".
Default = "only_Official".}
}
\value{
input data frame with master taxa information added to it.
}
\description{
This function has been deprecated (March 2026).
}
\details{
The new function is qc_taxa_match_official.

This function exists only as a wrapper to avoid breaking older code.
}
