% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa_values_character.R
\name{qc_taxa_values_char}
\alias{qc_taxa_values_char}
\title{QC Autecological Character Values}
\usage{
qc_taxa_values_char(data, col_char = NULL, valid_vals = NULL, separator = NULL)
}
\arguments{
\item{data}{A data frame containing autecological taxa data.}

\item{col_char}{The column containing the character values to be checked.}

\item{valid_vals}{Accepted values.}

\item{separator}{If values should be separated and checked include a
delimiter.  Default = NULL}
}
\value{
A data frame with col_char values, occurrence (n), and if valid (TRUE/
FALSE).  Any missing valid_vals are appended.
}
\description{
Performs basic QC of a character column against a list of accepted values.
}
\details{
Returns a data frame the values from the input with counts (column = n) from
the column and whether the values appeared in valid values (column =
valid). Values in the accepted values not appearing in the input are appended
to the bottom of the returned data frame.  These values are marked as n = NA
and valid = TRUE.  If NA is a valid value it must be included in valid_vals
or in the output NA will be labeled as valid = FALSE.

The default accepted values are the abbreviations are those used as
metric.values().

For Function Feeding Group (FFG); CF, CG, MH, OM, PA, PI, PR, SC, SH, and XY.  User using FC
and GC over CF and CG can modify the accepted values.  Both versions are
accepted in metric.values().

For Habit; BU, CB, CN, SK, SP, and SW.  Valid separated with "," are
first split apart and spaces removed.

life cycle

bcg_attr

habitat, habitat structure, elevation, gradient, thermal
}
\examples{
Values, FFG, Abr
qc_taxa_values_char(data_benthos_PacNW,
                    "FFG",
                    valid_vals = c("CF",
                                   "CG",
                                   "MH",
                                   "OM",
                                   "PA",
                                   "PH",
                                   "PI",
                                   "PR",
                                   "SC",
                                   "SH",
                                   "XY",
                                   NA))

# Values, FFG, full names
qc_taxa_values_char(data_benthos_MBSS,
                    "FFG",
                    valid_vals = c("Collector",
                                   "Filterer",
                                   "Predator",
                                   "Scraper",
                                   "Shredder"))

# Values, Habit, no separator
qc_taxa_values_char(data_benthos_MBSS,
                    "Habit",
                    valid_vals = c("bu", "cb", "cn", "dv", "sk", "sp", "sw"))

# Values, Habit, no separator
qc_taxa_values_char(data_benthos_MBSS,
                    "Habit",
                    valid_vals = c("bu", "cb", "cn", "dv", "sk", "sp", "sw"),
                    separator = ",")



}
