% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa_values_ffg.R
\name{qc_taxa_values_ffg}
\alias{qc_taxa_values_ffg}
\title{QC Functional Feeding Group (FFG) Values}
\usage{
qc_taxa_values_ffg(
  df_data,
  col_ffg,
  valid_vals = c("CF", "CG", "MH", "OM", "PA", "PH", "PI", "PR", "SC", "SH", "XY")
)
}
\arguments{
\item{df_data}{A data frame containing taxa data.}

\item{col_ffg}{The column containing FFG values (unquoted tidyselect style).}

\item{valid_vals}{Accepted values.
Default = c(CF, CG, MH, OM ,PA, PH, PI, PR, SC, SH, XY)}
}
\value{
A data frame with col_ffg values, occurrence (n), and if valid (TRUE/
FALSE).  Additional values from valid_vals are appended.
}
\description{
Performs basic QC of the FFG column against a list of accepted values.
}
\details{
Returns a data frame the values from the input with counts (column = n) from
the FFG column and whether the value appeared in valid values (column =
valid). Values in the accepted values not appearing in the input are appended
to the bottom of the returned data frame.  These values are marked as n = NA
and valid = TRUE.

The default accepted values are the abbreviations are those used as
metric.values(); CF, CG, MH, OM, PA, PI, PR, SC, SH, and XY.  User using FC
and GC over CF and CG can modify the accepted values.  Both versions are
accepted in metric.values().
}
\examples{
# Values, Default
qc_taxa_values_ffg(data_benthos_PacNW, "FFG")

# Values, User (full names)
qc_taxa_values_ffg(data_benthos_MBSS, "FFG",
                   valid_vals = c("Collector",
                                  "Filterer",
                                  "Predator",
                                  "Scraper",
                                  "Shredder"))

}
