% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa_values_habit.R
\name{qc_taxa_values_habit}
\alias{qc_taxa_values_habit}
\title{QC Habitat Values}
\usage{
qc_taxa_values_habit(
  df_data,
  col_habit = "Habit",
  valid_vals = c("BU", "CB", "CN", "SK", "SP", "SW")
)
}
\arguments{
\item{df_data}{A data frame containing taxa data.}

\item{col_habit}{The column containing Habit values.  Default = "Habit"}

\item{valid_vals}{Accepted values.
Default = c(BU, CB, CN, SK ,SP, SW.)}
}
\value{
A data frame with col_habit values, occurrence (n), and if valid
(TRUE/FALSE).  Additional values from valid_vals are appended.
}
\description{
Performs basic QC of the Habit column against a list of accepted values.
}
\details{
Returns a data frame the values from the input with counts (column = n) from
the Habit column and whether the value appeared in valid values (column =
valid). Values in the accepted values not appearing in the input are appended
to the bottom of the returned data frame.  These values are marked as n = NA
and valid = TRUE.

The default accepted values are the abbreviations are those used as
metric.values(); BU, CB, CN, SK, SP, and SW.  Valid separated with "," are
first split apart and spaces removed.
}
\examples{
# Values, Default
qc_taxa_values_habit(data_benthos_MBSS)

# Values, User
qc_taxa_values_habit(data_benthos_MBSS,
                     "Habit",
                     valid_vals = c("bu", "cb", "cn", "dv", "sk", "sp", "sw"))

}
