% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa_values_logical.R
\name{qc_taxa_values_logical}
\alias{qc_taxa_values_logical}
\title{QC Autecological Logical Values}
\usage{
qc_taxa_values_logical(data, col_logical = NULL)
}
\arguments{
\item{data}{A data frame containing autecological taxa data.}

\item{col_vals}{The column containing the logical value.}
}
\value{
A data frame with col_vals values, occurrence (n), and valid
(TRUE/FALSE).  Missing values (TRUE, FALSE, or NA) are appended.
}
\description{
Performs basic QC of a logical column showing occurrence.
}
\details{
Returns a data frame of the values from the input with counts (column = n) by
column.
}
\examples{
# Exclude
qc_taxa_values_logical(data_benthos_MBSS, "EXCLUDE")

# NonTarget
qc_taxa_values_logical(data_benthos_MBSS, "NONTARGET")

}
