% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa_values_tolval.R
\name{qc_taxa_values_tolval}
\alias{qc_taxa_values_tolval}
\title{QC Habitat Values}
\usage{
qc_taxa_values_tolval(
  df_data,
  col_tolval = "TolVal",
  valid_min = 0,
  valid_max = 10
)
}
\arguments{
\item{df_data}{A data frame containing taxa data.}

\item{col_tolval}{The column containing Tolerance Values.  Default = "TolVal"}

\item{valid_min}{Valid values range minimum.  Default = 0.}

\item{valid_max}{Valid values range maximum.  Default = 10.}
}
\value{
A data frame with col_tolval values, occurrence (n), and if valid
(TRUE/FALSE).
}
\description{
Performs basic QC of the Tolerance Value column.
}
\details{
Returns a data frame the values from the input with counts (column = n) from
the TolVal column and whether the value appeared in valid values (column =
valid).

The default accepted values are 0 - 10.
}
\examples{
qc_taxa_values_tolval(data_benthos_MBSS, "TOLVAL")

}
