% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_translate.R
\name{taxa_translate}
\alias{taxa_translate}
\title{Taxa Translate}
\usage{
taxa_translate(
  df_user = NULL,
  df_official = NULL,
  df_official_metadata = NULL,
  taxaid_user = "TAXAID",
  taxaid_official_match = NULL,
  taxaid_official_project = NULL,
  col_drop = NULL,
  sum_n_taxa_boo = FALSE,
  sum_n_taxa_col = NULL,
  sum_n_taxa_group_by = NULL
)
}
\arguments{
\item{df_user}{User taxa data}

\item{df_official}{Official project taxa data (master taxa list).}

\item{df_official_metadata}{Metadata for offiical project taxa data.
Included
Default is NULL}

\item{taxaid_user}{Taxonomic identifier in user data.  Default is "TAXAID".}

\item{taxaid_official_match}{Taxonomic identifier in official data user to
match with user data.  This is not the project taxanomic identifier.}

\item{taxaid_official_project}{Taxonomic identifier in official data that is
specific to a project, e.g., after operational taxonomic unit (OTU) applied.}

\item{col_drop}{Columns to remove in output.  Default = NULL}

\item{sum_n_taxa_boo}{Boolean value for if the results should be summarized
Default = FALSE}

\item{sum_n_taxa_col}{Column name for number of individuals for user data
when summarizing.  This column will be summed.
Default = NULL (suggestion = N_TAXA)}

\item{sum_n_taxa_group_by}{Column names for user data to use for grouping the
data when summarizing the user data.  Suggestions are SAMPID and TAXA_ID.
Default = NULL}
}
\value{
A list with three elements.  The first (merge) is the user data frame
with additional columns from the official data appended to it.  Names from
the user data that overlap with the official data have the suffix '_User'.
The second element (nonmatch) of the list is a vector of the non-matching
taxa from the user data.  The third element (metadata) includes the
metadata for the official data (if provided).
}
\description{
Convert user taxa names to those in an official project based
name list.
}
\details{
Merges user file with official file.  The official file has
phylogeny, autecology, and other project specific fields.

The inputs for the function uses existing data frames (or tibbles).

Any fields that match between the user file and the official file the
official data column name is retained.  The data fields from the user file
have the suffix "_USER".

By default, taxa are not collapsed to the official taxaid.  That is, if
multiple taxa in a sample have the same name the rows will not be combined.
If collapsing is desired set the parameter `sum_n_taxa_boo` to TRUE.
Will also need to provide `sum_n_taxa_col` and `sum_n_taxa_group_by`.

Slightly different than `qc_taxa` since no options in `taxa_translate` for
using one field over another and is more generic.
}
\examples{
None at this time
}
