% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxaMaster_Ben_BCG_PacNW}
\alias{TaxaMaster_Ben_BCG_PacNW}
\title{TaxaMaster_Ben_BCG_PacNW}
\format{
A data frame with 684 observations on the following 20 variables.

  \describe{
    \item{\code{TaxaID}}{a character vector}
    \item{\code{Phylum}}{a character vector}
    \item{\code{SubPhylum}}{a character vector}
    \item{\code{Class}}{a character vector}
    \item{\code{SubClass}}{a character vector}
    \item{\code{Order}}{a character vector}
    \item{\code{SuperFamily}}{a character vector}
    \item{\code{Family}}{a character vector}
    \item{\code{Tribe}}{a character vector}
    \item{\code{Genus}}{a character vector}
    \item{\code{SubGenus}}{a character vector}
    \item{\code{Species}}{a character vector}
    \item{\code{BCG_Attr}}{a character vector}
    \item{\code{NonTarget}}{a logical vector}
    \item{\code{Thermal_Indicator}}{a character vector}
    \item{\code{Long_Lived}}{a character vector}
    \item{\code{FFG}}{a character vector}
    \item{\code{Habit}}{a character vector}
    \item{\code{Life_Cycle}}{a character vector}
    \item{\code{TolVal}}{a numeric vector}
  }
}
\source{
example master taxa from BCG Pacific Northwest
}
\usage{
TaxaMaster_Ben_BCG_PacNW
}
\description{
Example data
}
\keyword{datasets}
