% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_diatom_mmi_dev}
\alias{data_diatom_mmi_dev}
\title{Diatom taxa data; Indiana DEM}
\format{

}
\source{
example data from IDEM
}
\usage{
data_diatom_mmi_dev
}
\description{
A data set with example diatom data.
Calculate metrics.  Data from IDEM.
}
\keyword{datasets}
