% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_diatom_mmi_qc}
\alias{data_diatom_mmi_qc}
\title{Diatom metric value data; Indiana DEM}
\format{
A data frame with 497 observations on the following 250 variables.

  \describe{
    \item{\code{SAMPLEID}}{a character vector}
    \item{\code{INDEX_NAME}}{a character vector}
    \item{\code{INDEX_REGION}}{a character vector}
    \item{\code{ni_total}}{a numeric vector}
    \item{\code{li_total}}{a numeric vector}
    \item{\code{nt_total}}{a numeric vector}
    \item{\code{nt_Achnan_Navic}}{a numeric vector}
    \item{\code{nt_LOW_N}}{a numeric vector}
    \item{\code{nt_HIGH_N}}{a numeric vector}
    \item{\code{nt_LOW_P}}{a numeric vector}
    \item{\code{nt_HIGH_P}}{a numeric vector}
    \item{\code{nt_BC_1}}{a numeric vector}
    \item{\code{nt_BC_2}}{a numeric vector}
    \item{\code{nt_BC_3}}{a numeric vector}
    \item{\code{nt_BC_4}}{a numeric vector}
    \item{\code{nt_BC_5}}{a numeric vector}
    \item{\code{nt_BC_12}}{a numeric vector}
    \item{\code{nt_BC_45}}{a numeric vector}
    \item{\code{nt_PT_1}}{a numeric vector}
    \item{\code{nt_PT_2}}{a numeric vector}
    \item{\code{nt_PT_3}}{a numeric vector}
    \item{\code{nt_PT_4}}{a numeric vector}
    \item{\code{nt_PT_5}}{a numeric vector}
    \item{\code{nt_PT_12}}{a numeric vector}
    \item{\code{nt_SALINITY_1}}{a numeric vector}
    \item{\code{nt_SALINITY_2}}{a numeric vector}
    \item{\code{nt_SALINITY_3}}{a numeric vector}
    \item{\code{nt_SALINITY_4}}{a numeric vector}
    \item{\code{nt_SALINITY_12}}{a numeric vector}
    \item{\code{nt_SALINITY_34}}{a numeric vector}
    \item{\code{nt_O_1}}{a numeric vector}
    \item{\code{nt_O_2}}{a numeric vector}
    \item{\code{nt_O_3}}{a numeric vector}
    \item{\code{nt_O_4}}{a numeric vector}
    \item{\code{nt_O_5}}{a numeric vector}
    \item{\code{nt_O_345}}{a numeric vector}
    \item{\code{nt_SESTONIC_HABIT}}{a numeric vector}
    \item{\code{nt_BENTHIC_HABIT}}{a numeric vector}
    \item{\code{nt_BAHLS_1}}{a numeric vector}
    \item{\code{nt_BAHLS_2}}{a numeric vector}
    \item{\code{nt_BAHLS_3}}{a numeric vector}
    \item{\code{nt_TROPHIC_1}}{a numeric vector}
    \item{\code{nt_TROPHIC_2}}{a numeric vector}
    \item{\code{nt_TROPHIC_3}}{a numeric vector}
    \item{\code{nt_TROPHIC_4}}{a numeric vector}
    \item{\code{nt_TROPHIC_5}}{a numeric vector}
    \item{\code{nt_TROPHIC_6}}{a numeric vector}
    \item{\code{nt_TROPHIC_7}}{a numeric vector}
    \item{\code{nt_TROPHIC_456}}{a numeric vector}
    \item{\code{nt_SAP_1}}{a numeric vector}
    \item{\code{nt_SAP_2}}{a numeric vector}
    \item{\code{nt_SAP_3}}{a numeric vector}
    \item{\code{nt_SAP_4}}{a numeric vector}
    \item{\code{nt_SAP_5}}{a numeric vector}
    \item{\code{nt_NON_N_FIXER}}{a numeric vector}
    \item{\code{nt_N_FIXER}}{a numeric vector}
    \item{\code{nt_HIGHLY_MOTILE}}{a numeric vector}
    \item{\code{nt_MODERATELY_MOTILE}}{a numeric vector}
    \item{\code{nt_NON_MOTILE}}{a numeric vector}
    \item{\code{nt_SLIGHTLY_MOTILE}}{a numeric vector}
    \item{\code{nt_WEAKLY_MOTILE}}{a numeric vector}
    \item{\code{nt_BIG}}{a numeric vector}
    \item{\code{nt_SMALL}}{a numeric vector}
    \item{\code{nt_MEDIUM}}{a numeric vector}
    \item{\code{nt_VERY_BIG}}{a numeric vector}
    \item{\code{nt_VERY_SMALL}}{a numeric vector}
    \item{\code{nt_ADNATE}}{a numeric vector}
    \item{\code{nt_STALKED}}{a numeric vector}
    \item{\code{nt_HIGHLY_MOTILE.1}}{a numeric vector}
    \item{\code{nt_ARAPHID}}{a numeric vector}
    \item{\code{nt_DIAT_CL_1}}{a numeric vector}
    \item{\code{nt_DIAT_CL_2}}{a numeric vector}
    \item{\code{nt_BEN_SES_1}}{a numeric vector}
    \item{\code{nt_BEN_SES_2}}{a numeric vector}
    \item{\code{nt_DIAT_CA_1}}{a numeric vector}
    \item{\code{nt_DIAT_CA_2}}{a numeric vector}
    \item{\code{nt_DIAT_COND_1}}{a numeric vector}
    \item{\code{nt_DIAT_COND_2}}{a numeric vector}
    \item{\code{nt_DIATAS_TN_1}}{a numeric vector}
    \item{\code{nt_DIATAS_TN_2}}{a numeric vector}
    \item{\code{nt_DIATAS_TP_1}}{a numeric vector}
    \item{\code{nt_DIATAS_TP_2}}{a numeric vector}
    \item{\code{nt_MOTILITY_1}}{a numeric vector}
    \item{\code{nt_MOTILITY_2}}{a numeric vector}
    \item{\code{nt_NF_1}}{a numeric vector}
    \item{\code{nt_NF_2}}{a numeric vector}
    \item{\code{pi_Achnan_Navic}}{a numeric vector}
    \item{\code{pi_HIGH_N}}{a numeric vector}
    \item{\code{pi_LOW_N}}{a numeric vector}
    \item{\code{pi_HIGH_P}}{a numeric vector}
    \item{\code{pi_LOW_P}}{a numeric vector}
    \item{\code{pi_BC_1}}{a numeric vector}
    \item{\code{pi_BC_2}}{a numeric vector}
    \item{\code{pi_BC_3}}{a numeric vector}
    \item{\code{pi_BC_4}}{a numeric vector}
    \item{\code{pi_BC_5}}{a numeric vector}
    \item{\code{pi_PT_1}}{a numeric vector}
    \item{\code{pi_PT_2}}{a numeric vector}
    \item{\code{pi_PT_3}}{a numeric vector}
    \item{\code{pi_PT_4}}{a numeric vector}
    \item{\code{pi_PT_5}}{a numeric vector}
    \item{\code{pi_PT_45}}{a numeric vector}
    \item{\code{pi_SALINITY_1}}{a numeric vector}
    \item{\code{pi_SALINITY_2}}{a numeric vector}
    \item{\code{pi_SALINITY_3}}{a numeric vector}
    \item{\code{pi_SALINITY_4}}{a numeric vector}
    \item{\code{pi_O_1}}{a numeric vector}
    \item{\code{pi_O_2}}{a numeric vector}
    \item{\code{pi_O_3}}{a numeric vector}
    \item{\code{pi_O_4}}{a numeric vector}
    \item{\code{pi_O_5}}{a numeric vector}
    \item{\code{pi_SESTONIC_HABIT}}{a numeric vector}
    \item{\code{pi_BENTHIC_HABIT}}{a numeric vector}
    \item{\code{pi_BAHLS_1}}{a numeric vector}
    \item{\code{pi_BAHLS_2}}{a numeric vector}
    \item{\code{pi_BAHLS_3}}{a numeric vector}
    \item{\code{pi_TROPHIC_1}}{a numeric vector}
    \item{\code{pi_TROPHIC_2}}{a numeric vector}
    \item{\code{pi_TROPHIC_3}}{a numeric vector}
    \item{\code{pi_TROPHIC_4}}{a numeric vector}
    \item{\code{pi_TROPHIC_5}}{a numeric vector}
    \item{\code{pi_TROPHIC_6}}{a numeric vector}
    \item{\code{pi_TROPHIC_7}}{a numeric vector}
    \item{\code{pi_SAP_1}}{a numeric vector}
    \item{\code{pi_SAP_2}}{a numeric vector}
    \item{\code{pi_SAP_3}}{a numeric vector}
    \item{\code{pi_SAP_4}}{a numeric vector}
    \item{\code{pi_SAP_5}}{a numeric vector}
    \item{\code{pi_NON_N_FIXER}}{a numeric vector}
    \item{\code{pi_N_FIXER}}{a numeric vector}
    \item{\code{pi_HIGHLY_MOTILE}}{a numeric vector}
    \item{\code{pi_MODERATELY_MOTILE}}{a numeric vector}
    \item{\code{pi_NON_MOTILE}}{a numeric vector}
    \item{\code{pi_SLIGHTLY_MOTILE}}{a numeric vector}
    \item{\code{pi_WEAKLY_MOTILE}}{a numeric vector}
    \item{\code{pi_BIG}}{a numeric vector}
    \item{\code{pi_SMALL}}{a numeric vector}
    \item{\code{pi_MEDIUM}}{a numeric vector}
    \item{\code{pi_VERY_BIG}}{a numeric vector}
    \item{\code{pi_VERY_SMALL}}{a numeric vector}
    \item{\code{pi_ADNATE}}{a numeric vector}
    \item{\code{pi_STALKED}}{a numeric vector}
    \item{\code{pi_HIGHLY_MOTILE.1}}{a numeric vector}
    \item{\code{pi_ARAPHID}}{a numeric vector}
    \item{\code{pi_DIAT_CL_1}}{a numeric vector}
    \item{\code{pi_DIAT_CL_1_ASSR}}{a numeric vector}
    \item{\code{pi_DIAT_CL_2}}{a numeric vector}
    \item{\code{pi_BEN_SES_1}}{a numeric vector}
    \item{\code{pi_BEN_SES_2}}{a numeric vector}
    \item{\code{pi_DIAT_CA_1}}{a numeric vector}
    \item{\code{pi_DIAT_CA_2}}{a numeric vector}
    \item{\code{pi_DIAT_COND_1}}{a numeric vector}
    \item{\code{pi_DIAT_COND_2}}{a numeric vector}
    \item{\code{pi_DIATAS_TN_1}}{a numeric vector}
    \item{\code{pi_DIATAS_TN_2}}{a numeric vector}
    \item{\code{pi_DIATAS_TP_1}}{a numeric vector}
    \item{\code{pi_DIATAS_TP_2}}{a numeric vector}
    \item{\code{pi_MOTILITY_1}}{a numeric vector}
    \item{\code{pi_MOTILITY_2}}{a numeric vector}
    \item{\code{pi_NF_1}}{a numeric vector}
    \item{\code{pi_NF_2}}{a numeric vector}
    \item{\code{pt_Achnan_Navic}}{a numeric vector}
    \item{\code{pt_HIGH_N}}{a numeric vector}
    \item{\code{pt_LOW_N}}{a numeric vector}
    \item{\code{pt_HIGH_P}}{a numeric vector}
    \item{\code{pt_LOW_P}}{a numeric vector}
    \item{\code{pt_BC_1}}{a numeric vector}
    \item{\code{pt_BC_2}}{a numeric vector}
    \item{\code{pt_BC_3}}{a numeric vector}
    \item{\code{pt_BC_4}}{a numeric vector}
    \item{\code{pt_BC_5}}{a numeric vector}
    \item{\code{pt_BC_12}}{a numeric vector}
    \item{\code{pt_BC_45}}{a numeric vector}
    \item{\code{pt_PT_1}}{a numeric vector}
    \item{\code{pt_PT_2}}{a numeric vector}
    \item{\code{pt_PT_3}}{a numeric vector}
    \item{\code{pt_PT_4}}{a numeric vector}
    \item{\code{pt_PT_5}}{a numeric vector}
    \item{\code{pt_PT_12}}{a numeric vector}
    \item{\code{pt_SALINITY_1}}{a numeric vector}
    \item{\code{pt_SALINITY_2}}{a numeric vector}
    \item{\code{pt_SALINITY_3}}{a numeric vector}
    \item{\code{pt_SALINITY_4}}{a numeric vector}
    \item{\code{pt_SALINITY_34}}{a numeric vector}
    \item{\code{pt_O_1}}{a numeric vector}
    \item{\code{pt_O_2}}{a numeric vector}
    \item{\code{pt_O_3}}{a numeric vector}
    \item{\code{pt_O_4}}{a numeric vector}
    \item{\code{pt_O_5}}{a numeric vector}
    \item{\code{pt_O_345}}{a numeric vector}
    \item{\code{pt_SESTONIC_HABIT}}{a numeric vector}
    \item{\code{pt_BENTHIC_HABIT}}{a numeric vector}
    \item{\code{pt_BAHLS_1}}{a numeric vector}
    \item{\code{pt_BAHLS_2}}{a numeric vector}
    \item{\code{pt_BAHLS_3}}{a numeric vector}
    \item{\code{pt_TROPHIC_1}}{a numeric vector}
    \item{\code{pt_TROPHIC_2}}{a numeric vector}
    \item{\code{pt_TROPHIC_3}}{a numeric vector}
    \item{\code{pt_TROPHIC_4}}{a numeric vector}
    \item{\code{pt_TROPHIC_5}}{a numeric vector}
    \item{\code{pt_TROPHIC_6}}{a numeric vector}
    \item{\code{pt_TROPHIC_7}}{a numeric vector}
    \item{\code{pt_TROPHIC_456}}{a numeric vector}
    \item{\code{pt_SAP_1}}{a numeric vector}
    \item{\code{pt_SAP_2}}{a numeric vector}
    \item{\code{pt_SAP_3}}{a numeric vector}
    \item{\code{pt_SAP_4}}{a numeric vector}
    \item{\code{pt_SAP_5}}{a numeric vector}
    \item{\code{pt_NON_N_FIXER}}{a numeric vector}
    \item{\code{pt_N_FIXER}}{a numeric vector}
    \item{\code{pt_HIGHLY_MOTILE}}{a numeric vector}
    \item{\code{pt_MODERATELY_MOTILE}}{a numeric vector}
    \item{\code{pt_NON_MOTILE}}{a numeric vector}
    \item{\code{pt_SLIGHTLY_MOTILE}}{a numeric vector}
    \item{\code{pt_WEAKLY_MOTILE}}{a numeric vector}
    \item{\code{pt_BIG}}{a numeric vector}
    \item{\code{pt_SMALL}}{a numeric vector}
    \item{\code{pt_MEDIUM}}{a numeric vector}
    \item{\code{pt_VERY_BIG}}{a numeric vector}
    \item{\code{pt_VERY_SMALL}}{a numeric vector}
    \item{\code{pt_ADNATE}}{a numeric vector}
    \item{\code{pt_STALKED}}{a numeric vector}
    \item{\code{pt_HIGHLY_MOTILE.1}}{a numeric vector}
    \item{\code{pt_ARAPHID}}{a numeric vector}
    \item{\code{pt_DIAT_CL_1}}{a numeric vector}
    \item{\code{pt_DIAT_CL_2}}{a numeric vector}
    \item{\code{pt_BEN_SES_1}}{a numeric vector}
    \item{\code{pt_BEN_SES_2}}{a numeric vector}
    \item{\code{pt_DIAT_CA_1}}{a numeric vector}
    \item{\code{pt_DIAT_CA_2}}{a numeric vector}
    \item{\code{pt_DIAT_COND_1}}{a numeric vector}
    \item{\code{pt_DIAT_COND_2}}{a numeric vector}
    \item{\code{pt_DIATAS_TN_1}}{a numeric vector}
    \item{\code{pt_DIATAS_TN_2}}{a numeric vector}
    \item{\code{pt_DIATAS_TP_1}}{a numeric vector}
    \item{\code{pt_DIATAS_TP_2}}{a numeric vector}
    \item{\code{pt_MOTILITY_1}}{a numeric vector}
    \item{\code{pt_MOTILITY_2}}{a numeric vector}
    \item{\code{pt_NF_1}}{a numeric vector}
    \item{\code{pt_NF_2}}{a numeric vector}
    \item{\code{nt_Sens_810}}{a numeric vector}
    \item{\code{nt_RefIndicators}}{a numeric vector}
    \item{\code{nt_Tol_13}}{a numeric vector}
    \item{\code{pi_Sens_810}}{a numeric vector}
    \item{\code{pi_RefIndicators}}{a numeric vector}
    \item{\code{pi_Tol_13}}{a numeric vector}
    \item{\code{pt_Sens_810}}{a numeric vector}
    \item{\code{pt_RefIndicators}}{a numeric vector}
    \item{\code{pt_Tol_13}}{a numeric vector}
    \item{\code{wa_POLL_TOL}}{a numeric vector}
  }
}
\source{
example metric value data from IDEM
}
\usage{
data_diatom_mmi_qc
}
\description{
A data set with diatom metric value data.
Used to compare to metric value calculations.  Data from IDEM.
}
\keyword{datasets}
