% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa_phylo.R
\name{qc_taxa_phylo}
\alias{qc_taxa_phylo}
\title{QC Taxa Phylo}
\usage{
qc_taxa_phylo(
  data,
  finalid = "FinalID",
  phylo_names = c("Phylum", "Subphylum", "Class", "Subclass", "Order", "Suborder",
    "Family", "Subfamily", "Tribe", "Genus"),
  ignore_case = FALSE
)
}
\arguments{
\item{data}{A data frame}

\item{finalid}{Column name for FinalID.  Default = "FinalID"}

\item{phylo_names}{Vector of phylogenetic names in order from coarse to fine.
Default = c("Phylum", "Subphylum", "Class", "Subclass", "Order", "Suborder",
"Family", "Subfamily", "Tribe", "Genus")}

\item{ignore_case}{Should case be ignored for checks.
Default = FALSE.}
}
\value{
A list with elements for each qc check.
}
\description{
Performs basic quality control on a phylogenetic list.
}
\details{
Returns a list of with elements corresponding the various checks on
a phylogenetic (standard or master) taxa list.

The phylogenetic list is multiple columns that the user will provide in rank
order from coarse to fine along with a FinalID column.

The checks are listed below and only report the entries that fail.

Some checks will detect those with potential issues but others that are
valid.

* **unique_parent** Each taxonomic rank (child) has a unique parent (coarser
rank).  Parents include all coarser ranks (as defined by user).

* **phylo_unique_rank** Each name is in only one phylogenetic rank column

* **phylo_as_finalid** Each phylogenetic name is also in FinalID

* **finalid_as_phylo** Each final id is a phylogenetic name

others checks?

case (all lower, all upper)

spaces
?
non A-Z (any case), e.g., slash, dash, underscore, parentheses, etc.

If ignore_case is TRUE then all columns (finalid and phylo_names) will be
converted to upper case before checks are performed.
}
\examples{
qc_phylo_PacNW <- qc_taxa_phylo(TaxaMaster_Ben_BCG_PacNW,
                                "TaxaID",
                                 phylo_names = c("Phylum",
                                                 "SubPhylum",
                                                 "Class",
                                                 "SubClass",
                                                 "Order",
                                                 "SuperFamily",
                                                 "Family",
                                                 "Tribe",
                                                 "Genus",
                                                 "SubGenus",
                                                 "Species"))
qc_phylo_PacNW$issues
qc_phylo_PacNW$unique_parent
qc_phylo_PacNW$phylo_unique_rank
qc_phylo_PacNW$phylo_as_finalid
qc_phylo_PacNW$finalid_as_phylo

}
