% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_taxa_values_numeric.R
\name{qc_taxa_values_numeric}
\alias{qc_taxa_values_numeric}
\title{QC Autecological Numeric Values}
\usage{
qc_taxa_values_numeric(
  data,
  col_numeric = NULL,
  valid_min = NULL,
  valid_max = NULL
)
}
\arguments{
\item{data}{A data frame containing autecological taxa data.}

\item{valid_min}{Valid values range minimum (inclusive).  Default = NA.}

\item{valid_max}{Valid values range maximum (inclusive).  Default = NA.}

\item{col_vals}{The column containing numeric values to be evaluated.
Default = NA.}
}
\value{
A data frame with col_vals values, occurrence (n), and valid
(TRUE/FALSE) within range of valid_min and valid_max.
}
\description{
Performs basic QC of a numeric column showing all values.
}
\details{
Returns a data frame of the values from the input with counts (column = n) from
the specified column.  User provided valid_min and valid_max are applied to
each set of values and evaluated as valid TRUE or FALSE.

The BioMonTools accepted values for TolVal are 0 - 10.

The BioMonTools accepted values for UFC are 1 - 6.
}
\examples{
# TolVal
qc_taxa_values_numeric(data_benthos_MBSS, "TOLVAL", 0, 10)

# TolVal2
qc_taxa_values_numeric(data_benthos_MBSS, "TOLVAL2", 0, 10)

# UFC
qc_taxa_values_numeric(data_benthos_MBSS, "UFC", 1, 6)

}
